requirements.txt
torch
torchvision
torchaudio
numpy
pandas
scikit-learn
tqdm
matplotlib
seaborn
Pillow
streamlit
opencv-python

